add also keys to openstack vim
