+++ catalogue/src/main/java/org/openbaton/catalogue/keys/PopKeypair.java
@@ -1,0 +1,7 @@
+package org.openbaton.catalogue.keys;
+
+import javax.persistence.Entity;
+import org.openbaton.catalogue.security.BaseKey;
+
+@Entity
+public class PopKeypair extends BaseKey {}
