+++ catalogue/src/main/java/org/openbaton/catalogue/security/BaseKey.java
@@ -1,0 +1,82 @@
+/*
+ * Copyright (c) 2016 Open Baton (http://www.openbaton.org)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.openbaton.catalogue.security;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Inheritance;
+import javax.persistence.InheritanceType;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
+import org.openbaton.catalogue.util.BaseEntity;
+
+@Entity
+@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
+public abstract class BaseKey extends BaseEntity {
+
+  @NotNull
+  @Size(min = 1)
+  private String name;
+
+  @NotNull
+  @Size(min = 1)
+  @Column(length = 500)
+  private String publicKey;
+
+  private String fingerprint;
+
+  @Override
+  public String toString() {
+    return "Key{"
+        + "name='"
+        + name
+        + '\''
+        + ", publicKey='"
+        + publicKey
+        + '\''
+        + ", fingerprint='"
+        + fingerprint
+        + '\''
+        + "} "
+        + super.toString();
+  }
+
+  public String getFingerprint() {
+    return fingerprint;
+  }
+
+  public void setFingerprint(String fingerprint) {
+    this.fingerprint = fingerprint;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public String getPublicKey() {
+    return publicKey;
+  }
+
+  public void setPublicKey(String publicKey) {
+    this.publicKey = publicKey;
+  }
+}
